refactor: connectionPairs to separate file
